API_DURABLE/requirements.txt
# =============================================================================
# LYA 6.0 - Requirements
# =============================================================================

# Azure Core
azure-functions>=1.21.0
azure-functions-durable>=1.2.9
azure-storage-blob

# AI APIs
google-genai>=1.0.0
anthropic>=0.40.0

# Document Processing
python-docx>=0.8.11
pdfplumber>=0.9.0

# NLP & ML (LYA 6.0)
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Utilities
requests
regex
tenacity
networkx>=3.2.1
python-dateutil
language-tool-python>=3.1.0